requirements.txt
<<<<<<< HEAD
cycler==0.10.0
decorator==4.4.0
imageio==2.5.0
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.3
numpy==1.17.0
opencv-contrib-python==4.1.0.25
pandas==0.25.0
Pillow==6.1.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytorch-ignite==0.2.0
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2
xlrd==1.2.0
=======
absl-py==0.7.1
albumentations==0.3.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
efficientnet-pytorch==0.4.0
entrypoints==0.3
et-xmlfile==1.0.1
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imgaug==0.2.6
importlib-metadata==0.18
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jdcal==1.4.1
jedi==0.14.0
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
more-itertools==7.1.0
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numpy==1.16.4
opencv-python==4.1.0.25
opencv-python-headless==4.1.0.25
openpyxl==2.6.2
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pathlib2==2.3.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
py==1.8.0
pydicom==1.3.0
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==4.6.3
python-dateutil==2.8.0
pytorch-ignite==0.2.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.5.1
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==1.2.0
torchvision==0.4.0
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.2
xlrd==1.2.0
zipp==0.5.1
>>>>>>> e9f5b35913e57866605023d79feffe82c7a528a3
